Refactored Comparator a bit.
